[REMOVED] Handler, all notify of HANDLER [ADDED] Normal realisation of ENUM

+++ src/main/java/Content/Boulder.java
@@ -15,7 +15,7 @@ public class Boulder extends Substance implements Exists {
     }
 
     @Override
-    public String toBe(Boolean negative, byte time) {
+    public String toBe(Boolean negative, int time) {
         String output = new String();
         if (negative & time == 0){
             output += "И есть ";
