+++ src/main/java/Interfaces/Exists.java
@@ -1,5 +1,5 @@
 package Interfaces;
 
 public interface Exists {
-    String toBe(Boolean negative, byte time);
+    String toBe(Boolean negative, int time);
 }
